use app.js for DefaultFixture
