--- packages/@aws-cdk-testing/cli-integ/resources/cdk-apps/from-lookup-dummy/cdk.json
@@ -1,7 +1,0 @@
-{
-  "app": "node app.js",
-  "versionReporting": false,
-  "context": {
-    "aws-cdk:enableDiffNoFail": "true"
-  }
-}
